security: Add flow offload API for OPI firewall
This commit adds a first cut at an OPI Security API for flow-offload,
with the goal being something which firewalls can use to offload flows.
This was initially taken from the sessionOffload project [1],
specifically the openoffload.proto [2]. I have modified this to make it
compile with the latest protobuf compilers, and fixed a few APIs to make
a bit more sense as well.

[1] https://github.com/att/sessionOffload
[2] https://github.com/att/sessionOffload/blob/master/protos/openoffload.proto

Closes #115

Signed-off-by: Kyle Mestery <mestery@mestery.com>