feat: (EVPN GW API) Add "components" field to each API object's status
This is an API enhancement for the EVPN GW use case.

The addition of a "components" field in the status sector of each
API object will enable the capture of the state and details
that are generated by the different subscribed components
(FRR component, Linux component, etc...) which are realizing the
configuration for each API object.

Signed-off-by: Dimitrios Markou <dimitrios.markou@ericsson.com>